requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0,<3.0
psycopg2-binary
python-jose
passlib[bcrypt]
bcrypt
pydantic[email]
jinja2
python-multipart
